+++ core/api_container/api_container_env_vars/api_container_env_vars.go
@@ -7,7 +7,9 @@ package api_container_env_vars
 
 // Constants for the environment variables that are used in the Dockerfile, made available to Go code
 const (
+	// TODO sort these alphabetically
 	TestSuiteContainerIdEnvVar = "TEST_SUITE_CONTAINER_ID"
+	TestVolumeName = "TEST_VOLUME"
 	NetworkIdEnvVar            = "NETWORK_ID"
 	SubnetMaskEnvVar           = "SUBNET_MASK"
 	GatewayIpEnvVar            = "GATEWAY_IP"